requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
faiss-cpu
langchain
pandas
langchain_google_genai
